Fix currentTarget property override in WebKit
WebKit 538 throws a TypeError when configuring the currentTarget property
and returns undefined for the property descriptor. So if the descriptor
is missing, avoid attempting to define the property.

We don't test the descriptor.configurable boolean because Safari 9
returns false despite it actually being configurable. WebKit nightly
builds correctly return true.

This affects PhantomJS 2.0 and Safari 8.0-8.0.5.